Move the version package to internal/
