+++ Dockerfile
@@ -12,6 +12,7 @@ RUN mkdir -p .git/objects
 COPY Makefile Makefile
 COPY OLM_VERSION OLM_VERSION
 COPY pkg pkg
+COPY internal internal
 COPY vendor vendor
 COPY go.mod go.mod
 COPY go.sum go.sum
