+++ pkg/lib/operatorstatus/status.go
@@ -20,8 +20,8 @@ import (
 	"k8s.io/client-go/discovery"
 
 	olmv1alpha1 "github.com/operator-framework/api/pkg/operators/v1alpha1"
+	olmversion "github.com/operator-framework/operator-lifecycle-manager/internal/version"
 	"github.com/operator-framework/operator-lifecycle-manager/pkg/lib/operatorclient"
-	olmversion "github.com/operator-framework/operator-lifecycle-manager/pkg/version"
 )
 
 const (
